support multiple community chips
