+++ apps/frontend/src/components/ui/HeaderCover.tsx
@@ -5,6 +5,7 @@ interface HeaderCoverProps {
   isLoading?: boolean;
   className?: string;
   size?: number;
+  image?: "edge-city" | "devcon";
 }
 
 export const CursiveLogo = ({
@@ -23,27 +24,50 @@ export const CursiveLogo = ({
   );
 };
 
-export const HeaderCover = ({ isLoading = false }: HeaderCoverProps) => {
+export const HeaderCover = ({ isLoading = false, image }: HeaderCoverProps) => {
   return (
     <div className="relative w-full top-0 pb-5">
       <div className="relative h-full">
-        <Image
-          src="/images/edge-city-logo.png"
-          alt="edge city logo"
-          className=" absolute left-4 top-4"
-          width={96}
-          height={57}
-        />
-        <Image
-          src="/images/register-main-cover-compressed.svg"
-          alt="register main"
-          className=" object-cover w-full"
-          width={100}
-          height={200}
-        />
+        {image === "edge-city" && (
+          <Image
+            src="/images/edge-city-logo.png"
+            alt="edge city logo"
+            className=" absolute left-4 top-4"
+            width={96}
+            height={57}
+          />
+        )}
+        {image === "edge-city" && (
+          <Image
+            src="/images/lanna_register_cover_compressed.svg"
+            alt="edge city register main"
+            className=" object-cover w-full"
+            width={100}
+            height={200}
+          />
+        )}
+        {image === "devcon" && (
+          <Image
+            src="/images/devcon_register_cover.svg"
+            alt="devcon register main"
+            className=" object-cover w-full"
+            width={100}
+            height={200}
+          />
+        )}
+        {image === undefined && (
+          <Image
+            src="/images/default_header_cover.svg"
+            alt="default cover"
+            className=" object-cover w-full"
+            width={100}
+            height={200}
+          />
+        )}
         <CursiveLogo
           className="position absolute -bottom-8 left-1/2 -ml-[30px]"
           isLoading={isLoading}
+          image={image}
         />
       </div>
     </div>
